extract command line UIs from *_gui.py
